refactor: importing config file consistency
